backend:refactor: split profile views
